refactor: split `useRecipeFiltering` into two distinct hooks with each of their concerns; one for (filter) selection, one for applying the (selected) filter
